requirements.txt
fastapi==0.86.0
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
